+++ srunner/scenarios/basic_scenario.py
@@ -17,23 +17,6 @@ from srunner.scenariomanager.scenario_manager import Scenario
 from srunner.scenariomanager.carla_data_provider import CarlaActorPool
 
 
-def get_location_in_distance(actor, distance):
-    """
-    Obtain a location in a given distance from the current actor's location.
-    Note: Search is stopped on first intersection.
-
-    @return obtained location and the traveled distance
-    """
-    waypoint = actor.get_world().get_map().get_waypoint(actor.get_location())
-    traveled_distance = 0
-    while not waypoint.is_intersection and traveled_distance < distance:
-        waypoint_new = waypoint.next(1.0)[-1]
-        traveled_distance += waypoint_new.transform.location.distance(waypoint.transform.location)
-        waypoint = waypoint_new
-
-    return waypoint.transform.location, traveled_distance
-
-
 class BasicScenario(object):
 
     """
