+++ srunner/scenarios/control_loss.py
@@ -20,6 +20,7 @@ from srunner.scenariomanager.atomic_scenario_behavior import *
 from srunner.scenariomanager.atomic_scenario_criteria import *
 from srunner.scenariomanager.timer import TimeOut
 from srunner.scenarios.basic_scenario import *
+from srunner.scenarios.scenario_helper import *
 
 CONTROL_LOSS_SCENARIOS = [
     "ControlLoss"
